Cleaned up comments

+++ src/edu/stanford/cs/gaming/sdk/model/ObjProperty.java
@@ -7,7 +7,8 @@ public class ObjProperty {
      public String name;
      public int prop_type;
      public int int_val;
-     public float float_val;
+//     public float float_val;
+     public double float_val;
      public String string_val;
      public byte[] blob_val;
